backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
Pillow>=10.0
google-genai>=1.0.0
python-dotenv>=1.0
